refactor: rename timeline actions
