+++ web/src/lib/components/timeline/AssetSelectControlBar.svelte
