+++ web/src/lib/components/timeline/actions/SelectAllAction.svelte
